+++ core/banking/oneoff_transfers.go
@@ -86,7 +86,7 @@ func (e *Engine) oneOffTransfer(
 		return err
 	}
 
-	if err := e.ensureMinimalTransferAmount(a, transfer.Amount); err != nil {
+	if err := e.ensureMinimalTransferAmount(a, transfer.Amount, transfer.FromAccountType, transfer.From); err != nil {
 		transfer.Status = types.TransferStatusRejected
 		return err
 	}