+++ core/banking/transfer_common_test.go
@@ -81,3 +81,49 @@ func TestCheckTransfer(t *testing.T) {
 	transfer.Amount = num.NewUint(200)
 	require.EqualError(t, e.CheckTransfer(transfer), "could not transfer funds, not enough funds to transfer")
 }
+
+func TestCheckTransferWithVestedAccount(t *testing.T) {
+	e := getTestEngine(t)
+
+	transfer := &types.TransferBase{
+		From:            "03ae90688632c649c4beab6040ff5bd04dbde8efbf737d8673bbda792a110301",
+		FromAccountType: types.AccountTypeVestedRewards,
+		To:              "03ae90688632c649c4beab6040ff5bd04dbde8efbf737d8673bbda792a110301",
+		ToAccountType:   types.AccountTypeGeneral,
+		Asset:           "eth",
+		Amount:          num.NewUint(10),
+		Reference:       "someref",
+	}
+
+	e.OnMinTransferQuantumMultiple(context.Background(), num.DecimalFromFloat(1))
+
+	// balance is under the min amount
+	e.col.EXPECT().GetPartyVestedRewardAccount(gomock.Any(), gomock.Any()).Return(&types.Account{Balance: num.NewUint(90)}, nil).Times(1)
+
+	// asset exists
+	e.assets.EXPECT().Get(gomock.Any()).Times(1).Return(assets.NewAsset(&mockAsset{num.DecimalFromFloat(100)}), nil)
+	// try to transfer a small balance, but not the whole balance
+	require.EqualError(t,
+		e.CheckTransfer(transfer),
+		"transfer from vested account under minimal transfer amount must be the full balance",
+	)
+
+	// now we try to transfre the full amount
+	e.col.EXPECT().GetPartyVestedRewardAccount(gomock.Any(), gomock.Any()).Return(&types.Account{Balance: num.NewUint(90)}, nil).Times(2)
+	transfer.Amount = num.NewUint(90)
+	e.assets.EXPECT().Get(gomock.Any()).Times(1).Return(assets.NewAsset(&mockAsset{num.DecimalFromFloat(100)}), nil)
+	require.NoError(t,
+		e.CheckTransfer(transfer),
+	)
+
+	// now we try again, with a balance above the min amount, but not the whole balance
+
+	e.col.EXPECT().GetPartyVestedRewardAccount(gomock.Any(), gomock.Any()).Return(&types.Account{Balance: num.NewUint(300)}, nil).Times(1)
+	e.assets.EXPECT().Get(gomock.Any()).Times(1).Return(assets.NewAsset(&mockAsset{num.DecimalFromFloat(100)}), nil)
+
+	transfer.Amount = num.NewUint(110)
+	// try to transfer a small balance, but not the whole balance
+	require.NoError(t,
+		e.CheckTransfer(transfer),
+	)
+}